+++ app/controllers/users/registrations_controller.rb
@@ -2,6 +2,14 @@
 
 module Users
   class RegistrationsController < Devise::RegistrationsController
+    before_action :check_captcha, only: %i[create] # rubocop: disable Rails/LexicallyScopedActionFilter
+
+    private
+
+    def check_captcha
+      redirect_to new_user_registration_path unless verify_recaptcha
+    end
+
     protected
 
     def after_inactive_sign_up_path_for(resource)