+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.1.1] - 20.03.2021
 ### Added
 - sitemap
 - Freedom Finance broker importing data
 - rendering list of operations
 - captcha for users registration
+- vue js support
 
 ### Modified
 - use cache for storing quote coupons sum for time range